+++ libsonuma/sonuma_daemon.c
@@ -152,7 +152,7 @@ int kal_reg_ctx(int fd, uint8_t **ctx_ptr, uint32_t num_pages)
 }
 
 /* Msutherl: New version of rmc_send, using paired send/recv slots */
-void rmc_send(rmc_wq_t *wq, char *lbuff_ptr, int lbuff_offset, size_t size, int snid, uint8_t sending_qp, send_slot_t* send_slot,uint8_t slot_idx)
+void rmc_send(rmc_wq_t *wq, char *lbuff_ptr, int lbuff_offset, size_t size, int snid, uint8_t sending_qp, send_slot_t* send_slot,uint16_t slot_idx)
 {
     uint8_t wq_head = wq->head;
     DLog("[rmc_send] Entering rmc_send.");
@@ -192,7 +192,7 @@ void rmc_send(rmc_wq_t *wq, char *lbuff_ptr, int lbuff_offset, size_t size, int 
     }
 }
 
-void rmc_recv(rmc_wq_t *wq, char* lbuff_ptr,int lbuff_offset,size_t size,int snid,uint8_t sending_qp,uint8_t slot_idx)
+void rmc_recv(rmc_wq_t *wq, char* lbuff_ptr,int lbuff_offset,size_t size,int snid,uint16_t sending_qp,uint16_t slot_idx)
 {
     // create WQ entry, response for arguments given to CQ
     uint8_t wq_head = wq->head;
